Revamp completion popup layout, refactor completion logic, and update styling and similarity calculation

+++ hextant-core/src/main/kotlin/hextant/completion/CompletionResult.kt
@@ -20,6 +20,6 @@ sealed class CompletionResult {
      */
     data class Match(
         val matchedRegions: List<IntRange>,
-        val similarity: Int
+        val similarity: Double
     ) : CompletionResult()
 }
